chore: rename js assets to ts
