+++ src/index.html
@@ -63,6 +63,6 @@
         </div>
     </div>
 
-    <script type="module" src="js/app.ts"></script>
+    <script type="module" src="ts/app.ts"></script>
 </body>
 </html>
